requirements.txt
anthropic>=0.96
pydantic>=2,<3
pillow
mujoco>=3
robosuite>=1.5
robomimic==0.3.0
imageio
imageio-ffmpeg
numpy
torch
h5py
ruff
mypy
pytest
pytest-asyncio
